Readability smoothing (Line 6)
From [link](https://www.merriam-webster.com/words-at-play/words-to-not-begin-sentences-with):

"The Merriam-Webster Dictionary of English Usage begins its entry on and with this statement: 'Everybody agrees that it’s all right to begin a sentence with and, and nearly everybody admits to having been taught at some past time that the practice was wrong.' The entry notes that there has been speculation that sentence-initial ands were discouraged to prevent children from stringing together interminable lists of clauses or sentences. While it makes sense to avoid an enormous pile of independent clauses in one's writing, it seems unlikely that trying to change the way people naturally use the language will solve this problem. There are times when it would be ill-advised to begin a sentence with and or but, and there are times when it works just fine."

While not technically wrong, I believe compounding these two sentences with a comma in this instance increases the flow of the writing and leads to increased readability for the user.